refactor(Toast): modularize Toast component into separate files
This commit restructures the Toast component by splitting it into smaller, more maintainable modules: Toast, ToastIcon, ToastContainer, ToastContent, ToastService, and types. This improves code readability, reusability, and separation of concerns. The functionality remains unchanged, but the codebase is now better organized for future enhancements.
