+++ src/components/Toast/ToastContainer.tsx
@@ -1,0 +1,30 @@
+import type { Component } from "solid-js";
+import { Portal } from "solid-js/web";
+import type { ToastContainerProps } from "./types";
+
+/**
+ * Toast container component - Manages toast position and animation effects
+ */
+const ToastContainer: Component<ToastContainerProps> = (props) => {
+  return (
+    <Portal
+      ref={(el) => {
+        el.classList.add("fluent-toast-container");
+      }}
+    >
+      <div
+        class={`fluent-toast fluent-toast-${props.position}`}
+        style={{
+          ...props.style,
+          "z-index": props.zIndex,
+        }}
+        role="alert"
+        aria-live="assertive"
+      >
+        {props.children}
+      </div>
+    </Portal>
+  );
+};
+
+export default ToastContainer;
