+++ src/components/Toast/index.scss
@@ -78,7 +78,6 @@
       box-shadow: 0 4px 12px rgba(0, 0, 0, 0.3);
     }
 
-    // 不同类型的Toast样式
     &.fluent-toast-info {
       border-left-color: var(--colorBrandBackground, #0078d4);
       .fluent-toast-icon {
